Update deepcopy to handle nested documents
In queries that include nested documents, deepcopy wasn't
capturing that the nested documents were documents,
which is sometimes tested using isinstance. This ensures
that the type is retained through deep copies.

This adds two tests, as well. "without - 38" failed prior to
this patch. "without - 39" succeeded, but was included to
test that the order of the merge() and without() methods was
signficant in triggering the deepcopy issue.